+++ sortbootorder.c
@@ -52,12 +52,12 @@ static int fetch_file_from_cbfs( char *filename, char destination[MAX_DEVICES][M
 static int get_line_number(u8 line_start, u8 line_end, char key );
 static void int_ids( char buffer[MAX_DEVICES][MAX_LENGTH], u8 line_cnt, u8 lineDef_cnt );
 static void save_flash(char buffer[MAX_DEVICES][MAX_LENGTH], u8 max_lines);
-static void update_tag_value(char buffer[MAX_DEVICES][MAX_LENGTH], u8 max_lines, const char * tag, char value);
+static void update_tag_value(char buffer[MAX_DEVICES][MAX_LENGTH], u8 *max_lines, const char * tag, char value);
 
 /*** local variables ***/
 static u8 ipxe_toggle;
-static u8 serial_toggle;
 static u8 usb_toggle;
+static u8 sga_toggle;
 static char bootlist_def[MAX_DEVICES][MAX_LENGTH];
 static char bootlist_map[MAX_DEVICES][MAX_LENGTH];
 static char id[MAX_DEVICES] = {0};
@@ -75,6 +75,7 @@ static u8 device_toggle[MAX_DEVICES];
  *        - Serial console disable / enable
  *        - Network / IPXE disable / enable
  *        - USB boot disable / enable
+ *        - SgaBios disable / enable
  *        - Exit with or without saving order
  */
 
@@ -87,9 +88,7 @@ int main(void) {
 	u8 bootlist_map_ln = 0;
 	u8 line_start = 0;
 	u8 line_number = 0;
-	char *ipxe_str;
-	char *scon_str;
-	char *usb_str;
+	char *token;
 	struct cbfs_handle *bootorder_handle;
 
 	// Set to enabled because enable toggle is not (yet) implemented for these devices
@@ -116,17 +115,17 @@ int main(void) {
 	fetch_file_from_cbfs( BOOTORDER_MAP, bootlist_map, &bootlist_map_ln );
 
 	// Init ipxe and serial status
-	ipxe_str = cbfs_find_string("pxen", BOOTORDER_FILE);
-	ipxe_str += strlen("pxen");
-	ipxe_toggle = ipxe_str ? strtoul(ipxe_str, NULL, 10) : 1;
+	token = cbfs_find_string("pxen", BOOTORDER_FILE);
+	token += strlen("pxen");
+	ipxe_toggle = token ? strtoul(token, NULL, 10) : 1;
 
-	scon_str = cbfs_find_string("scon", BOOTORDER_FILE);
-	scon_str += strlen("scon");
-	serial_toggle = scon_str ? strtoul(scon_str, NULL, 10) : 1;
+	token = cbfs_find_string("usben", BOOTORDER_FILE);
+	token += strlen("usben");
+	usb_toggle = token ? strtoul(token, NULL, 10) : 1;
 
-	usb_str = cbfs_find_string("usben", BOOTORDER_FILE);
-	usb_str += strlen("usben");
-	usb_toggle = usb_str ? strtoul(usb_str, NULL, 10) : 1;
+	token = cbfs_find_string("sgaen", BOOTORDER_FILE);
+	token += strlen("sgaen");
+	sga_toggle = token ? strtoul(token, NULL, 10) : 0;
 
 	show_boot_device_list( bootlist, max_lines, bootlist_def_ln );
 	int_ids( bootlist, max_lines, bootlist_def_ln );
@@ -142,10 +141,6 @@ int main(void) {
 					copy_list_line(&(bootlist_def[i][0]), &(bootlist[i][0]));
 				int_ids( bootlist, max_lines, bootlist_def_ln );
 				break;
-			case 't':
-			case 'T':
-				serial_toggle ^= 0x1;
-				break;
 			case 'n':
 			case 'N':
 				ipxe_toggle ^= 0x1;
@@ -154,11 +149,15 @@ int main(void) {
 			case 'U':
 				usb_toggle ^= 0x1;
 				break;
+			case 'l':
+			case 'L':
+				sga_toggle ^= 0x1;
+				break;
 			case 's':
 			case 'S':
-				update_tag_value(bootlist, max_lines, "scon", serial_toggle + '0');
-				update_tag_value(bootlist, max_lines, "pxen", ipxe_toggle + '0');
-				update_tag_value(bootlist, max_lines, "usben", usb_toggle + '0');
+				update_tag_value(bootlist, &max_lines, "pxen", ipxe_toggle + '0');
+				update_tag_value(bootlist, &max_lines, "usben", usb_toggle + '0');
+				update_tag_value(bootlist, &max_lines, "sgaen", sga_toggle + '0');
 				save_flash( bootlist, max_lines );
 				// fall through to exit ...
 			case 'x':
@@ -248,7 +247,7 @@ static void show_boot_device_list( char buffer[MAX_DEVICES][MAX_LENGTH], u8 line
 	printf("\n\n");
 	printf("  r Restore boot order defaults\n");
 	printf("  n Network/PXE boot - Currently %s\n", (ipxe_toggle) ? "Enabled" : "Disabled");
-	printf("  t Serial console - Currently %s\n", (serial_toggle) ? "Enabled" : "Disabled");
+	printf("  l Serial console redirection - Currently %s\n", (sga_toggle) ? "Enabled" : "Disabled");
 	printf("  u USB boot - Currently %s\n", (usb_toggle) ? "Enabled" : "Disabled");
 	printf("  x Exit setup without save\n");
 	printf("  s Save configuration and exit\n");
@@ -293,15 +292,15 @@ static int fetch_file_from_cbfs( char *filename, char destination[MAX_DEVICES][M
 		else
 			char_cnt++;
 
-		if ( *line_count > MAX_DEVICES) {
+		if (*line_count > MAX_DEVICES) {
 			printf("aborting due to excessive line_count\n");
 			break;
 		}
-		if ( char_cnt > MAX_LENGTH) {
+		if (char_cnt > MAX_LENGTH) {
 			printf("aborting due to excessive char count\n");
 			break;
 		}
-		if ( cbfs_offset > (MAX_LENGTH*MAX_DEVICES)) {
+		if (cbfs_offset > (MAX_LENGTH*MAX_DEVICES)) {
 			printf("aborting due to excessive cbfs ptr length\n");
 			break;
 		}
@@ -376,14 +375,28 @@ static void save_flash(char buffer[MAX_DEVICES][MAX_LENGTH], u8 max_lines) {
 }
 
 /*******************************************************************************/
-static void update_tag_value(char buffer[MAX_DEVICES][MAX_LENGTH], u8 max_lines, const char * tag, char value)
+static void update_tag_value(char buffer[MAX_DEVICES][MAX_LENGTH], u8 *max_lines, const char * tag, char value)
 {
 	int i;
+	bool found = FALSE;
 
-	for (i = 0; i < max_lines; i++) {
+	for (i = 0; i < *max_lines; i++) {
 		if (!strncmp(tag, &buffer[i][0], strlen(tag))) {
+			found = TRUE;
 			buffer[i][strlen(tag)] = value;
 			break;
 		}
+	}
+
+	if (!found) {
+		if ((*max_lines + 1) > MAX_DEVICES) {
+			return;
+		}
+		strcpy(&buffer[*max_lines][0], tag);
+		buffer[*max_lines][strlen(tag)] = value;
+		buffer[*max_lines][strlen(tag)+1] = '\r';
+		buffer[*max_lines][strlen(tag)+2] = '\n';
+		buffer[*max_lines][strlen(tag)+3] = '0';
+		(*max_lines)++;
 	}
 }